LayoutTests/storage/domstorage/sessionstorage/simple-usage.html

<html>
<head>
<script src="resources/clearSessionStorage.js"></script>
<script>

if (window.layoutTestController)
    layoutTestController.dumpAsText();

function log(a)
{
    document.getElementById("logger").innerHTML += a + "<br>";
}

function runTest()
{
    if (!window.sessionStorage) {
        log("window.sessionStorage DOES NOT exist");
        return;
    }
    
    log("Length is " + window.sessionStorage.length);
    log("Value for FOO is " + window.sessionStorage.getItem("FOO"));

    window.sessionStorage.setItem("FOO", "BAR");
    
    log("Length is " + window.sessionStorage.length);
    log("Value for FOO is " + window.sessionStorage.getItem("FOO"));
    log("Key for index 0 is " + window.sessionStorage.key(0));
    try {
        window.sessionStorage.key(1);
    } catch(e) {
        log("Correctly caught exception trying to access key at index 1");
        log(e);
    }
    
    window.sessionStorage.setItem("FOO", "BAZ");
    
    log("Length is " + window.sessionStorage.length);
    log("Value for FOO is " + window.sessionStorage.getItem("FOO"));
    
    window.sessionStorage.removeItem("FOO");
    
    log("Length is " + window.sessionStorage.length);
    log("Value for FOO is " + window.sessionStorage.getItem("FOO"));
}

</script>
</head>
<body onload="runTest();">
This test trys simple operations on SessionStorage<br>
<div id="logger"></div>
</body>
</html>
