third_party/blink/web_tests/platform/generic/storage/domstorage/sessionstorage/simple-usage-expected.txt

This test trys simple operations on SessionStorage

Length is 0
Value for FOO is null
Length is 1
Value for FOO is BAR
Key for index 0 is FOO
Key for index 1 is null
Key for index -1 is null
Length is 1
Value for FOO is BAZ
Length is 0
Value for FOO is null
